Removed unused code #355
