+++ CareKit/CareKit/View Updaters/TaskCategory/OCKDetailedTaskCategoryView+Updatable.swift
@@ -1,4 +1,3 @@
-//
 /*
  Copyright (c) 2020, Apple Inc. All rights reserved.
