+++ CareKitStore/CareKitStoreTests/Structs/TestTaskCategoryQuery.swift
@@ -1,4 +1,3 @@
-//
 /*
  Copyright (c) 2020, Apple Inc. All rights reserved.
