+++ CareKitUI/CareKitUI/Common/Views/OCKHeaderView.swift
@@ -1,21 +1,21 @@
 /*
  Copyright (c) 2019, Apple Inc. All rights reserved.
- 
+
  Redistribution and use in source and binary forms, with or without modification,
  are permitted provided that the following conditions are met:
- 
+
  1.  Redistributions of source code must retain the above copyright notice, this
  list of conditions and the following disclaimer.
- 
+
  2.  Redistributions in binary form must reproduce the above copyright notice,
  this list of conditions and the following disclaimer in the documentation and/or
  other materials provided with the distribution.
- 
+
  3. Neither the name of the copyright holder(s) nor the names of any contributors
  may be used to endorse or promote products derived from this software without
  specific prior written permission. No license is granted to the trademarks of
  the copyright holders even if such marks are included in this software.
- 
+
  THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -59,6 +59,9 @@ open class OCKHeaderView: OCKView {
 
         /// Flag to show an image on the leading side of the text in the view.
         public var showsIconImage: Bool = false
+
+        /// Flag to indicate if the icon image should be cropped to a circular shape
+        public var isIconImageCircular: Bool = true
     }
 
     // MARK: Properties
@@ -127,7 +130,7 @@ open class OCKHeaderView: OCKView {
         configurationHandler(&configuration)
         self.configuration = configuration
 
-        iconImageView = configuration.showsIconImage ? OCKHeaderView.makeIconImageView() : nil
+        iconImageView = configuration.showsIconImage ? OCKHeaderView.makeIconImageView(isIconImageCircular: configuration.isIconImageCircular) : nil
         detailDisclosureImage = configuration.showsDetailDisclosure ? OCKHeaderView.makeDetailDisclosureImage() : nil
         separatorView = configuration.showsSeparator ? OCKSeparatorView() : nil
         super.init()
@@ -174,9 +177,25 @@ open class OCKHeaderView: OCKView {
         contentStackView.setCustomSpacing(margin, after: headerStackView)
     }
 
-    private static func makeIconImageView() -> UIImageView {
-        let imageView = OCKCircleImageView()
+    /// Multi-line detail label below `titleLabel`.
+    private static func makeDetailLabel() -> OCKLabel {
+        let label = OCKLabel(textStyle: .caption1, weight: .medium)
+
+        label.numberOfLines = 0
+        label.animatesTextChanges = true
+
+        return label
+    }
+
+    private static func makeIconImageView(isIconImageCircular: Bool) -> UIImageView {
+        let imageView: UIImageView
+        if isIconImageCircular {
+            imageView = OCKCircleImageView()
+        } else {
+            imageView = UIImageView()
+        }
         imageView.contentMode = .scaleAspectFill
+
         return imageView
     }
